requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-experimental
sqlalchemy
faiss-cpu
sentence-transformers
openai
